fix unknowns possibly?

+++ model.py
@@ -118,7 +118,7 @@ class PixelCNN(nn.Module):
 
         ### ADDED BY AN IDIOT (ME) ###
 
-        embeddings = self.label_embeddings(labels)
+        embeddings = self.label_embeddings(labels.long())
 
         g,b = self.filmer(embeddings)
